moving fixtures into conftest.py
